Renaming single_space_after_token to whitespace_after_token.
