+++ vsg/rules/__init__.py
@@ -37,7 +37,6 @@ from .align_consecutive_lines_after_line_starting_with_token_and_stopping_with_t
 from .whitespace_before_token import whitespace_before_token
 from .whitespace_before_tokens_in_between_tokens import whitespace_before_tokens_in_between_tokens
 from .remove_blank_lines_above_line_starting_with_token import remove_blank_lines_above_line_starting_with_token
-from .single_space_after_token import single_space_after_token
 from .single_space_before_token import single_space_before_token
 from .single_space_before_token_if_on_same_line_as_token import single_space_before_token_if_on_same_line_as_token
 from .blank_lines_between_token_pairs import blank_lines_between_token_pairs
