+++ vsg/rules/alias_declaration/rule_102.py
@@ -1,7 +1,7 @@
 
 from vsg import token
 
-from vsg.rules import single_space_after_token as Rule
+from vsg.rules.whitespace_after_token import Rule as Rule
 
 lTokens = []
 lTokens.append(token.alias_declaration.is_keyword)
