+++ vsg/rules/attribute_declaration/rule_100.py
@@ -1,14 +1,14 @@
 
 from vsg import token
 
-from vsg.rules import single_space_after_token
+from vsg.rules.whitespace_after_token import Rule
 
 lTokens = []
 lTokens.append(token.attribute_declaration.attribute_keyword)
 lTokens.append(token.attribute_declaration.colon)
 
 
-class rule_100(single_space_after_token):
+class rule_100(Rule):
     '''
     This rule checks for a single space after the following elements:  **attribute** keyword and colon.
 
@@ -25,4 +25,4 @@ class rule_100(single_space_after_token):
        attribute max_delay : time;
     '''
     def __init__(self):
-        single_space_after_token.__init__(self, 'attribute_declaration', '100', lTokens)
+        Rule.__init__(self, 'attribute_declaration', '100', lTokens)
