+++ vsg/rules/attribute_specification/rule_100.py
@@ -1,7 +1,7 @@
 
 from vsg import token
 
-from vsg.rules import single_space_after_token
+from vsg.rules.whitespace_after_token import Rule
 
 lTokens = []
 lTokens.append(token.attribute_specification.attribute_keyword)
@@ -10,7 +10,7 @@ lTokens.append(token.attribute_specification.of_keyword)
 lTokens.append(token.attribute_specification.is_keyword)
 
 
-class rule_100(single_space_after_token):
+class rule_100(Rule):
     '''
     This rule checks for a single space after the following attribute_specification elements:  **attribute** keyword, *attribute_designator*, **of** keyword and **is** keyword.
 
@@ -31,4 +31,4 @@ class rule_100(single_space_after_token):
        attribute coordinate of comp_1:component   is (0.0, 17.5);
     '''
     def __init__(self):
-        single_space_after_token.__init__(self, 'attribute_specification', '100', lTokens)
+        Rule.__init__(self, 'attribute_specification', '100', lTokens)
