+++ vsg/rules/concurrent/rule_002.py
@@ -1,5 +1,5 @@
 
-from vsg.rules import single_space_after_token
+from vsg.rules.whitespace_after_token import Rule
 
 from vsg import token
 
@@ -9,7 +9,7 @@ lTokens.append(token.concurrent_conditional_signal_assignment.assignment)
 lTokens.append(token.concurrent_selected_signal_assignment.assignment)
 
 
-class rule_002(single_space_after_token):
+class rule_002(Rule):
     '''
     This rule checks for a single space after the **<=** operator.
 
@@ -29,4 +29,4 @@ class rule_002(single_space_after_token):
     '''
 
     def __init__(self):
-        single_space_after_token.__init__(self, 'concurrent', '002', lTokens)
+        Rule.__init__(self, 'concurrent', '002', lTokens)
