+++ vsg/rules/entity_specification/rule_100.py
@@ -1,13 +1,13 @@
 
 from vsg import token
 
-from vsg.rules import single_space_after_token
+from vsg.rules.whitespace_after_token import Rule
 
 lTokens = []
 lTokens.append(token.entity_specification.colon)
 
 
-class rule_100(single_space_after_token):
+class rule_100(Rule):
     '''
     This rule checks for a single space after the colon.
 
@@ -26,4 +26,4 @@ class rule_100(single_space_after_token):
        attribute coordinate of comp_1 : component is (0.0, 17.5);
     '''
     def __init__(self):
-        single_space_after_token.__init__(self, 'entity_specification', '100', lTokens)
+        Rule.__init__(self, 'entity_specification', '100', lTokens)
