+++ vsg/rules/if_statement/rule_005.py
@@ -2,7 +2,7 @@
 from vsg import parser
 from vsg import token
 
-from vsg.rules import single_space_after_token as Rule
+from vsg.rules.whitespace_after_token import Rule as Rule
 
 lTokens = []
 lTokens.append(token.if_statement.elsif_keyword)
