+++ vsg/rules/signal/rule_005.py
@@ -1,13 +1,13 @@
 
 from vsg import token
 
-from vsg.rules import single_space_after_token
+from vsg.rules.whitespace_after_token import Rule
 
 lTokens = []
 lTokens.append(token.signal_declaration.colon)
 
 
-class rule_005(single_space_after_token):
+class rule_005(Rule):
     '''
     This rule checks for a single space after the colon.
 
@@ -26,5 +26,5 @@ class rule_005(single_space_after_token):
        signal rd_en : std_logic;
     '''
     def __init__(self):
-        single_space_after_token.__init__(self, 'signal', '005', lTokens)
+        Rule.__init__(self, 'signal', '005', lTokens)
         self.solution = 'Ensure only a signal space after the colon.'
