+++ vsg/rules/variable/rule_005.py
@@ -1,13 +1,13 @@
 
 from vsg import token
 
-from vsg.rules import single_space_after_token
+from vsg.rules.whitespace_after_token import Rule
 
 lTokens = []
 lTokens.append(token.variable_declaration.colon)
 
 
-class rule_005(single_space_after_token):
+class rule_005(Rule):
     '''
     This rule checks there is a single space after the colon.
 
@@ -26,5 +26,5 @@ class rule_005(single_space_after_token):
        variable counter : integer;
     '''
     def __init__(self):
-        single_space_after_token.__init__(self, 'variable', '005', lTokens)
+        Rule.__init__(self, 'variable', '005', lTokens)
         self.solution = 'Ensure only a single space after the colon.'
